[EC/VDP] Adding handicap friendly parkings

+++ src/main/java/Day3/UsageNotification.java
@@ -2,7 +2,7 @@ package Day3;
 
 public class UsageNotification {
 
-    public Object sendNotification(int capacity) {
+    public Object sendNotification(double capacity) {
         return String.format("Parking Usage is at %d%%", capacity);
     }
 }
